rename to shorter RemoveVoidDocblockFromMagicMethodsRector
