--- rules-tests/DeadCode/Rector/ClassMethod/RemoveUselessVoidReturnFromDocblockOnVoidMagicMethodsRector/Fixture/remove_return_construct.php.inc
@@ -1,30 +1,0 @@
-<?php
-
-namespace Rector\Tests\DeadCode\Rector\ClassMethod\RemoveUselessVoidReturnFromDocblockOnVoidMagicMethodsRector\Fixture;
-
-class RemoveReturnConstruct
-{
-    /**
-     * @return void
-     */
-    function __construct()
-    {
-        //
-    }
-}
-
-?>
------
-<?php
-
-namespace Rector\Tests\DeadCode\Rector\ClassMethod\RemoveUselessVoidReturnFromDocblockOnVoidMagicMethodsRector\Fixture;
-
-class RemoveReturnConstruct
-{
-    function __construct()
-    {
-        //
-    }
-}
-
-?>
