--- rules-tests/DeadCode/Rector/ClassMethod/RemoveUselessVoidReturnFromDocblockOnVoidMagicMethodsRector/Fixture/remove_return_destruct.php.inc
@@ -1,30 +1,0 @@
-<?php
-
-namespace Rector\Tests\DeadCode\Rector\ClassMethod\RemoveUselessVoidReturnFromDocblockOnVoidMagicMethodsRector\Fixture;
-
-class RemoveReturnDestruct
-{
-    /**
-     * @return void
-     */
-    function __destruct()
-    {
-        //
-    }
-}
-
-?>
------
-<?php
-
-namespace Rector\Tests\DeadCode\Rector\ClassMethod\RemoveUselessVoidReturnFromDocblockOnVoidMagicMethodsRector\Fixture;
-
-class RemoveReturnDestruct
-{
-    function __destruct()
-    {
-        //
-    }
-}
-
-?>
