+++ rules-tests/DeadCode/Rector/ClassMethod/RemoveVoidDocblockFromMagicMethodsRector/Fixture/remove_return_construct.php.inc
@@ -1,0 +1,30 @@
+<?php
+
+namespace Rector\Tests\DeadCode\Rector\ClassMethod\RemoveVoidDocblockFromMagicMethodsRector\Fixture;
+
+class RemoveReturnConstruct
+{
+    /**
+     * @return void
+     */
+    function __construct()
+    {
+        //
+    }
+}
+
+?>
+-----
+<?php
+
+namespace Rector\Tests\DeadCode\Rector\ClassMethod\RemoveVoidDocblockFromMagicMethodsRector\Fixture;
+
+class RemoveReturnConstruct
+{
+    function __construct()
+    {
+        //
+    }
+}
+
+?>
