+++ rules-tests/DeadCode/Rector/ClassMethod/RemoveVoidDocblockFromMagicMethodsRector/Fixture/remove_return_destruct.php.inc
@@ -1,0 +1,30 @@
+<?php
+
+namespace Rector\Tests\DeadCode\Rector\ClassMethod\RemoveVoidDocblockFromMagicMethodsRector\Fixture;
+
+class RemoveReturnDestruct
+{
+    /**
+     * @return void
+     */
+    function __destruct()
+    {
+        //
+    }
+}
+
+?>
+-----
+<?php
+
+namespace Rector\Tests\DeadCode\Rector\ClassMethod\RemoveVoidDocblockFromMagicMethodsRector\Fixture;
+
+class RemoveReturnDestruct
+{
+    function __destruct()
+    {
+        //
+    }
+}
+
+?>
