+++ rules-tests/DeadCode/Rector/ClassMethod/RemoveVoidDocblockFromMagicMethodsRector/config/configured_rule.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+use Rector\Config\RectorConfig;
+use Rector\DeadCode\Rector\ClassMethod\RemoveVoidDocblockFromMagicMethodsRector;
+
+return RectorConfig::configure()
+    ->withRules([RemoveVoidDocblockFromMagicMethodsRector::class]);
